// client/src/DxDbe/src/services/ToolbarPanelManager/ButtonNames.ts
/*******************************************************************************
 * (C) Copyright
 * Logivations GmbH, Munich 2010-2022
 ******************************************************************************/

enum ButtonNames {
	reloadButton = 'reloadButton',
	executeTableUpdateButton = 'executeTableUpdateButton',
	getLinkButton = 'getLinkButton',
	toggleSelectionButton = 'toggleSelectionButton',
	deleteRowsButton = 'deleteRowsButton',
	addRowButton = 'addRowButton',
	cloneRowsButton = 'cloneRowsButton',
	saveRowsButton = 'saveRowsButton',
	saveForeignKeyButton = 'saveForeignKeyButton',
	switchViewButton = 'switchViewButton',
	massUpdateButton = 'massUpdateButton',
	exportButton = 'exportButton',
	importButton = 'importButton',
	settingsButton = 'settingsButton',
	pagerPanel = 'pagerPanel',
	closeSplitPanel = 'closeSplitPanel',
	compareButton = 'compareButton',
	masterButton = 'masterButton',
	helpButton = 'helpButton',
	linksMenuButton = 'linksMenuButton',
	tableActionsMenuButton = 'tableActionsMenuButton',
	filterSettingSplitMenuButton = 'filterSettingSplitMenuButton',
	hiddenColumnSettingSplitMenuButton = 'hiddenColumnSettingSplitMenuButton',
	groupingSettingSplitMenuButton = 'groupingSettingSplitMenuButton',
	tableSortButton = 'tableSortButton',
	childRecordsMenuButton = 'childRecordsMenuButton',
	parentRecordsMenuButton = 'parentRecordsMenuButton',
	viewChartMenuButton = 'viewChartMenuButton',
	synchronizeChildButton = 'synchronizeChildButton',
	mergeButton = 'mergeButton',
	screenBuilderButton = 'screenBuilderButton',
	columnChooserButton = 'columnChooserButton',
	linkReferenceButton = 'linkReferenceButton',
	closeScreenBuilderTable = 'closeScreenBuilderTable',
	rowCounterRefreshButton = 'rowCounterRefreshButton',
	goToOldDBE = 'goToOldDBE',
}

export default ButtonNames;